pyproject.toml
[project]
name = "qwen35-scope"
version = "0.1.0"
description = "SAE-based behavioral decomposition for Qwen 3.5-27B hybrid architecture"
requires-python = ">=3.11"
dependencies = [
    "torch>=2.4",
    "transformers>=4.48",
    "safetensors>=0.4",
    "einops>=0.7",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "anthropic>=0.40",
    "wandb>=0.18",
    "matplotlib>=3.8",
    "plotly>=5.18",
    "pandas>=2.1",
    "numpy>=1.26",
    "tqdm>=4.66",
    "datasets>=2.16",
    "jsonlines>=4.0",
    "scikit-learn>=1.4",
    "scipy>=1.12",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "ipython", "jupyter", "ruff"]
gpu = ["flash-attn>=2.5"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
testpaths = ["tests"]
